flights-search-agent/requirements.txt
# A2A Protocol SDK
a2a-sdk[http-server]>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.2
pydantic-settings>=2.6.0

# LangChain Core
langchain>=0.3.7
langchain-core>=0.3.15
langchain-community>=0.3.5

# AWS Bedrock Integration
boto3>=1.35.36
botocore>=1.35.36
langchain-aws>=0.2.6

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.2
